increase timeout for media uploads
